requirements.txt
apache-airflow-providers-apache-spark==5.3.0
pytrends
google-genai==1.19.0
google-generativeai==0.8.5
pandas
pyarrow
elasticsearch